2023-05-08 like 기능 완성

+++ src/main/java/com/kh/learnovation/domain/freeboard/controlller/LikeController.java
@@ -3,11 +3,14 @@ package com.kh.learnovation.domain.freeboard.controlller;
 
 import com.kh.learnovation.domain.freeboard.dto.LikeDTO;
 import com.kh.learnovation.domain.freeboard.service.LikeService;
-import com.kh.learnovation.domain.freeboard.service.LikeServiceImpl;
+import com.kh.learnovation.domain.user.dto.UserDTO;
+import com.kh.learnovation.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 
 @RestController
@@ -16,19 +19,27 @@ import org.springframework.web.bind.annotation.*;
 public class LikeController {
 
     private final LikeService likeService;
+    private final UserService userService;
 
 
-    @PostMapping("/check")
+    @PostMapping("/like")
     public ResponseEntity check(@ModelAttribute LikeDTO likeDTO){
-        int result =likeService.check(likeDTO);
-        if(result==0){
-            likeService.save(likeDTO);
-            int count = 1;
-            return new ResponseEntity<>(count, HttpStatus.OK);
+        Optional<UserDTO> optionalUserDTO = userService.getCurrentUser();
+        if (optionalUserDTO.isPresent()){
+            UserDTO userDTO =optionalUserDTO.get();
+            likeDTO.setUserId(userDTO.getId());
+            int result =likeService.check(likeDTO);
+            if(result==0){
+                likeService.save(likeDTO);
+                return new ResponseEntity<>("liked", HttpStatus.OK);
+            }else {
+                likeService.delete(likeDTO);
+                int count = 0;
+                return new ResponseEntity<>("unliked", HttpStatus.OK);
+            }
         }else {
-            likeService.delete(likeDTO);
-            int count = 0;
-            return new ResponseEntity<>(count, HttpStatus.OK);
+            return new ResponseEntity<>("실패", HttpStatus.OK);
+
         }
     }
